Fixed: Issue in EditalbleTextComponent of underflow of the UInt32 representing the cursor position.  This resulted in code trying to remove indexes out of bounds in the text.  This was fixed by checking if the cursor position is 0.

+++ Source/Contrib/UserInterface/src/Component/Text/OSGEditableTextComponent.cpp
@@ -115,8 +115,11 @@ void EditableTextComponent::keyTyped(const KeyEventUnrecPtr e)
 			{	
                 //erase at the current caret position
                 Int32 DeleteIndex(getCaretPosition());
-                moveCaret(-1);
-                deleteRange(DeleteIndex-1, DeleteIndex);
+                if(DeleteIndex != 0)
+                {
+                    moveCaret(-1);
+                    deleteRange(DeleteIndex-1, DeleteIndex);
+                }
 			}
 		}
 		if(e->getKey()== e->KEY_DELETE)
